+++ packages/multsig-wallet/src/index.ts
@@ -1,1 +1,10 @@
-export * from  "./multisig-wallet";
+/**
+ * @license MIT
+ * @copyright (c) 2022 Dirt Road Dev
+ * @package @skaleproject/multisig-wallet
+ *
+ * @file index.ts
+ * @author Sawyer Cutler
+*/
+
+export * from  "./contract";
